git-svn-id: file:///srv/local/data/cleanRepo/trunk@15211 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/sergioo/src/main/java/playground/sergioo/GTFS/Trip.java
@@ -12,7 +12,7 @@ public class Trip {
 	//Attributes
 	private Service service;
 	private Shape shape;
-	private SortedMap<Integer,StopTime> stopTimes; 
+	private SortedMap<Double,StopTime> stopTimes; 
 	private List<Frequency> frequencies;
 	private List<Link> links;
 	private Route route;
@@ -26,7 +26,7 @@ public class Trip {
 		super();
 		this.service = service;
 		this.shape = shape;
-		this.stopTimes = new TreeMap<Integer, StopTime>();
+		this.stopTimes = new TreeMap<Double, StopTime>();
 		this.frequencies = new ArrayList<Frequency>();
 		this.links = new ArrayList<Link>();
 		this.route = route;
@@ -46,7 +46,7 @@ public class Trip {
 	/**
 	 * @return the stopTimes
 	 */
-	public SortedMap<Integer, StopTime> getStopTimes() {
+	public SortedMap<Double, StopTime> getStopTimes() {
 		return stopTimes;
 	}
 	/**
@@ -72,7 +72,7 @@ public class Trip {
 	 * @param key
 	 * @param stopTime
 	 */
-	public void putStopTime(Integer key, StopTime stopTime) {
+	public void putStopTime(Double key, StopTime stopTime) {
 		stopTimes.put(key, stopTime);
 	}
 	/**
